--- app/dashboard/page.css
@@ -1,50 +1,0 @@
-.dashboard {
-    font-family: Arial, sans-serif;
-    margin: 20px;
-}
-
-.dashboard h1 {
-    font-size: 24px;
-    margin-bottom: 10px;
-}
-
-.dashboard h2 {
-    font-size: 18px;
-    margin-top: 20px;
-}
-
-.dashboard ul {
-    list-style-type: none;
-    margin: 0;
-    padding: 0;
-}
-
-.dashboard li {
-    margin-bottom: 10px;
-}
-
-.dashboard form {
-    margin-top: 20px;
-}
-
-.dashboard input[type="text"],
-.dashboard input[type="number"] {
-    padding: 5px;
-    margin-right: 10px;
-}
-
-.dashboard button[type="submit"] {
-    padding: 5px 10px;
-    background-color: #007bff;
-    color: #fff;
-    border: none;
-    cursor: pointer;
-}
-
-.dashboard button[type="submit"]:hover {
-    background-color: #0056b3;
-}
-
-p {
-    margin-bottom: 1em;
-}
