--- app/login/page.css
@@ -1,22 +1,0 @@
-.container {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    margin-top: 50px;
-    width: 600px;
-}
-
-.login-title {
-    text-align: center;
-    margin-bottom: 20px;
-    font-size: 24px;
-    font-weight: bold;
-}
-
-.form-group {
-    margin-bottom: 20px;
-}
-
-p {
-    margin-bottom: 1em;
-}
