--- app/page.css
@@ -1,32 +1,0 @@
-body {
-    font-family: Arial, sans-serif;
-    margin: 0;
-    padding: 0;
-    background-color: #1f1f1f;
-    color: #ddd;
-}
-
-.container {
-    max-width: 800px;
-    margin: 50px auto;
-    padding: 20px;
-    background-color: #333;
-    border-radius: 5px;
-    box-shadow: 0 0 10px rgba(255, 255, 255, 0.1);
-}
-
-h1, p {
-    color: #ddd;
-}
-
-.code-box {
-    background-color: #222;
-    border: 1px solid #444;
-    border-radius: 5px;
-    padding: 10px;
-    margin-top: 20px;
-}
-
-.code-box p {
-    color: #ddd;
-}
