+++ next.config.mjs
@@ -1,6 +1,4 @@
 /** @type {import('next').NextConfig} */
-const nextConfig = {
-    output: 'standalone',
-};
+const nextConfig = {};
 
 export default nextConfig;
